refactor(core)!: pipeline refactor (#37)
This change both moves the logic of pipeline building and creation to
the relevant classes, and also separates the render and compute
pipelines into separate classes. This leads to a much more organized
codebase, and also allows for easier extension of the pipeline system
later on.

BREAKING CHANGE: Pipeline creation has changed. Instead of
differentiating between render and compute pipelines via a type
parameter, they are now separate classes. This means that to create
pipelines you will just call the respective constructor.